% Generated by build script
\name{ensembl_sequence}
\alias{ensembl_sequence}
\title{Get gene sequence}
\usage{
ensembl_sequence()
}
\arguments{
\item{id}{Ensembl ID}
\item{type}{"genomic" (default), "cdna", "cds", "protein"}
}
\value{tibble: one row with id, seq (character), molecule_type}
\description{Get gene sequence}
