% Generated by build script
\name{ensembl_variant}
\alias{ensembl_variant}
\title{Look up variant by rsID}
\usage{
ensembl_variant()
}
\arguments{
\item{rsid}{Variant ID (e.g. "rs699")}
\item{species}{Species (default "homo_sapiens")}
}
\value{tibble: id, consequence_type, alleles, minor_allele, maf}
\description{Look up variant by rsID}
